refactor: move ad client to separate package
