+++ ad/__init__.py
@@ -1,0 +1,8 @@
+from .ad_client import ADUser, search_candidates, reset_password, disable_user
+
+__all__ = [
+    "ADUser",
+    "search_candidates",
+    "reset_password",
+    "disable_user",
+]
