+++ ad/ad_client.py
